[Hooks] Extract common hooks to utils
