+++ rest-tests/src/test/java/io/syndesis/qe/RestTestHooks.java
@@ -1,7 +1,5 @@
 package io.syndesis.qe;
 
-import static org.junit.Assume.assumeTrue;
-
 import io.syndesis.qe.bdd.storage.StepsStorage;
 import io.syndesis.qe.bdd.validation.OperatorValidationSteps;
 import io.syndesis.qe.endpoints.TestSupport;
@@ -10,18 +8,12 @@ import io.syndesis.qe.resource.impl.Jaeger;
 import io.syndesis.qe.utils.OpenShiftUtils;
 import io.syndesis.qe.utils.RestUtils;
 import io.syndesis.qe.utils.SampleDbConnectionManager;
-import io.syndesis.qe.utils.TestUtils;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
-import java.util.List;
 import java.util.concurrent.TimeoutException;
-import java.util.stream.Collectors;
 
-import cucumber.api.Scenario;
 import cucumber.api.java.After;
-import cucumber.api.java.Before;
-import io.fabric8.kubernetes.api.model.Pod;
 import lombok.extern.slf4j.Slf4j;
 
 /**
@@ -34,12 +26,6 @@ public class RestTestHooks {
 
     @Autowired
     private StepsStorage stepStorage;
-
-    @Before("@prod")
-    public void skipProdForNightly() {
-        // Skip prod tests when not running with productized build
-        assumeTrue(TestConfiguration.syndesisVersion().contains("redhat"));
-    }
 
     @After("@sqs")
     public void clearSqsProperty() {
@@ -61,24 +47,6 @@ public class RestTestHooks {
         OpenShiftUtils.getInstance().persistentVolumes().list().getItems().stream()
             .filter(pv -> pv.getMetadata().getName().startsWith(OperatorValidationSteps.TEST_PV_NAME))
             .forEach(pv -> OpenShiftUtils.getInstance().persistentVolumes().withName(pv.getMetadata().getName()).cascading(true).delete());
-    }
-
-    @After
-    public void getLogs(Scenario scenario) {
-        if (scenario.isFailed()) {
-            TestUtils.printPods();
-            log.warn("Scenario {} failed, saving integration logs to scenario", scenario.getName());
-            // There can be multiple integration pods for one test
-            List<Pod> integrationPods = OpenShiftUtils.getInstance().pods().list().getItems().stream().filter(
-                p -> p.getMetadata().getName().startsWith("i-")
-                    && !p.getMetadata().getName().contains("deploy")
-                    && !p.getMetadata().getName().contains("build")
-            ).collect(Collectors.toList());
-            for (Pod integrationPod : integrationPods) {
-                scenario.embed(String.format("%s\n\n%s", integrationPod.getMetadata().getName(),
-                    OpenShiftUtils.getInstance().getPodLog(integrationPod)).getBytes(), "text/plain");
-            }
-        }
     }
 
     @After("@publicapi-connections")
